refactor: remove old "gestión de productos" implementation and related components
Deleted legacy `ProductList`, `ProductoForm`, and associated `categorias.ts`. Removed unused dependencies and large redundant code blocks to clean up the codebase, aligning with the new inventory module.
