--- src/app/configuracion/productos/types/categorias.ts
@@ -1,5 +1,0 @@
-export interface Categoria {
-  id: string;
-  nombre: string;
-  color: string;
-}
