+++ src/components/GestionInventario/table/InventarioMobileList.tsx
@@ -19,6 +19,7 @@ import { IProductoTiendaV2 } from "@/schemas/producto";
 import { formatCurrency, formatNumber } from "@/utils/formatters";
 import { useAppContext } from "@/context/AppContext";
 import { getRentabilidad } from "./rentabilidad";
+import { generateProductCodesPDF } from "@/utils/productCodesPdf";
 
 interface Props {
   productos: IProductoTiendaV2[];
@@ -147,6 +148,17 @@ function ProductCard({
               }}
             >
               Historial movimientos
+            </MenuItem>
+            <MenuItem
+              onClick={async () => {
+                setAnchor(null);
+                await generateProductCodesPDF(
+                  p.producto.nombre,
+                  p.producto.codigosProducto,
+                );
+              }}
+            >
+              Descargar códigos PDF
             </MenuItem>
             <MenuItem
               onClick={() => {
